legacy build
fixes
