+++ packages/playground/src/app.tsx
@@ -4,8 +4,6 @@ import { eventBus, setEventBus } from './utils/serviceWorker';
 import { Update } from './components/update';
 import { useZoom } from 'solid-repl/src/hooks/useZoom';
 import { Edit } from './pages/edit';
-import { Home } from './pages/home';
-import { Login } from './pages/login';
 import { AppContextProvider } from './context';
 
 export const App = (): JSX.Element => {
@@ -38,10 +36,7 @@ export const App = (): JSX.Element => {
           </AppContextProvider>
         )}
       >
-        <Route path={['/:user/:repl', '/scratchpad']} component={Edit} />
-        <Route path="/:user" component={Home} />
-        <Route path="/" component={Home} />
-        <Route path="/login" component={Login} />
+        <Route path={['/:user/:repl', '/']} component={Edit} />
       </Router>
 
       <Show when={eventBus()} children={<Update onDismiss={() => setEventBus(false)} />} />
