+++ packages/playground/src/context.tsx
@@ -1,10 +1,8 @@
-import { Accessor, createContext, createResource, createSignal, ParentComponent, Resource, useContext } from 'solid-js';
+import { Accessor, createContext, createSignal, ParentComponent, useContext } from 'solid-js';
 import type { Tab } from 'solid-repl';
 import { isDarkTheme } from './utils/isDarkTheme';
 
 interface AppContextType {
-  token: string;
-  user: Resource<{ display: string; avatar: string } | undefined>;
   tabs: Accessor<Tab[] | undefined>;
   setTabs: (x: Accessor<Tab[] | undefined> | undefined) => void;
   dark: Accessor<boolean>;
@@ -18,25 +16,6 @@ const AppContext = createContext<AppContextType>();
 export const API = 'https://api.solidjs.com';
 
 export const AppContextProvider: ParentComponent = (props) => {
-  const [token, setToken] = createSignal(localStorage.getItem('token') || '');
-  const [user] = createResource(token, async (token) => {
-    if (!token)
-      return {
-        display: '',
-        avatar: '',
-      };
-    const result = await fetch(`${API}/profile`, {
-      headers: {
-        authorization: `Bearer ${token}`,
-      },
-    });
-    const body = await result.json();
-    return {
-      display: body.display,
-      avatar: body.avatar,
-    };
-  });
-
   const [dark, setDark] = createSignal(isDarkTheme());
   document.body.classList.toggle('dark', dark());
 
@@ -44,14 +23,6 @@ export const AppContextProvider: ParentComponent = (props) => {
   return (
     <AppContext.Provider
       value={{
-        get token() {
-          return token();
-        },
-        set token(x) {
-          setToken(x);
-          localStorage.setItem('token', x);
-        },
-        user,
         tabs() {
           const tabs = tabsGetter();
           if (!tabs) return undefined;
